Fix docker-compose.yaml volume bindings
The current volume definitions didn't work for me. In particular, $PWD didn't work, and it tried to bind a directory at `/etc/caddy/Caddyfile/` (which failed because `./Caddyfile` is a file not a dir).

This configuration works for me running docker in WSL2, but I'm unsure if this works just as well in other environments.